delete image resolved

+++ routes/movie.js
@@ -49,13 +49,7 @@ router.get('/', verify, (req, res) => {
             let userMovie=[];
             for( let i =0; i< movie.length; i++){
                 if(movie[i].owner === req.user._id){
-                    if(!fs.existsSync('.'+movie[i].img)){
-                        movie[i].img = '/public/defaultimage.jpg'
-                        console.log("in If ", movie[i].img)
-                        Movie.findById(movie[i]._id, {img: '/public/defaultimage.jpg'}, {new :true}).then(movie => {
-                            console.log("in updateById", movie._id)
-                        }).catch(err => { console.log(err)})
-                    }
+                   
                     userMovie.push(movie[i]);
                 }
             }
@@ -88,14 +82,14 @@ router.put('/update/:id', verify, async (req, res) => {
             let imageDirectory;
             if(req.files){
                 const movie = await Movie.findById(id);
-                if(movie.img !== '' && movie.img !== '/public/defaultimage.jpg'){
+                if(movie.img !== ''){
                     const path = '.'+movie.img;
-                    fs.unlink(path, (err) => {
+                    if(fs.existsSync(path)){fs.unlink(path, (err) => {
                         if (err) {
                         console.error(err)
                         return res.status(400).send(err);
                         }
-                    })
+                    })}
                         
                     Movie.findByIdAndUpdate(id, {img : ''})
                     .then( () => {
@@ -140,7 +134,7 @@ router.delete('/delete/:id', verify, async(req, res) => {
     try{
     const movie = await Movie.findById(id);
     const imageDirectory = movie.img;
-    if(imageDirectory !== '' && imageDirectory !== '/public/defaultimage.jpg'){
+    if(imageDirectory !== ''){
         if(fs.existsSync('.'+imageDirectory)){
         fs.unlinkSync('.'+imageDirectory);}
     }
